pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shruggery"
version = "0.1.0"
description = "Local Atlassian MCP server — full Jira Cloud v3 + Confluence Cloud v2 coverage with attachment handling"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
shruggery = "shruggery.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/shruggery"]
